Biblioteca BMP

+++ 2025/L_Ricci/src/lib/BMP085/BMP085.cpp
@@ -108,7 +108,7 @@ float BMP085::readTemperature() {
   return celsius;
 }
 
-float BMP085::readPressure() {
+long BMP085::readPressure() {
   return pressao;
 }
